+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [update] Update package json.
 - [refactor] Add lint and prettier.
 - [refactor] Castle and Magic.
+- [doc] Update readme file.
 
 # v0.4.0